Refactor Distributed helpers (#78)
* Refactor Distributed helpers

* fix

* Review suggestion
