+++ src/lightly_train/_modules/teachers/dinov2/dinov2_helper.py
@@ -7,43 +7,17 @@
 from __future__ import annotations
 
 import logging
-import os
 from pathlib import Path
 
 import torch
 
+from lightly_train import _distributed as distributed_helpers
 from lightly_train._data.download import download_from_url
 from lightly_train._modules.teachers.dinov2.models.vision_transformer import (
     DinoVisionTransformer,
 )
 
 logger = logging.getLogger(__name__)
-
-
-def get_local_rank() -> int | None:
-    """Get the local rank of the current process."""
-    rank_keys = ("LOCAL_RANK", "SLURM_LOCALID", "JSM_NAMESPACE_LOCAL_RANK")
-    for key in rank_keys:
-        rank = os.environ.get(key)
-        if rank is not None:
-            return int(rank)
-    return None
-
-
-def get_node_rank() -> int | None:
-    """Get the node rank of the current process."""
-    rank_keys = ("NODE_RANK", "GROUP_RANK", "SLURM_NODEID")
-    for key in rank_keys:
-        rank = os.environ.get(key)
-        if rank is not None:
-            return int(rank)
-    return None
-
-
-def is_local_rank_zero() -> bool:
-    """Check if the current process is running on the local rank zero."""
-    local_rank = get_local_rank()
-    return local_rank == 0 or local_rank is None
 
 
 def load_weights(
@@ -54,7 +28,7 @@ def load_weights(
 
     # Cache the teacher checkpoint. concatenate the node rank to the checkpoint path
     # to avoid overwriting the checkpoint if multiple nodes are used.
-    node_rank = get_node_rank()
+    node_rank = distributed_helpers.get_node_rank()
     if node_rank is not None:
         file_name = f"{str(node_rank)}_{str(Path(url).name)}"
     else:
@@ -62,7 +36,7 @@ def load_weights(
     checkpoint_path = checkpoint_dir / Path(file_name)
 
     # Only the global rank zero downloads the checkpoint.
-    if is_local_rank_zero():
+    if distributed_helpers.is_local_rank_zero():
         if not checkpoint_path.exists():
             logger.info(
                 f"Downloading teacher weights from: '{url}' and saving them to: "
